1.3.0内测版本1

+++ src/main/java/me/xuxiaoxiao/chatapi/wechat/protocol/RspInit.java
@@ -79,7 +79,7 @@ public class RspInit {
         public long OwnerUin;
         public int MemberCount;
         public ArrayList<User> MemberList;
-        public int Statues;
+        public long Statues;
         public long AttrStatus;
         public long MemberStatus;
         public String Province;
